requirements.txt
setuptools>=68.0.0
pyyaml==6.0.2
sphinx==7.1.2
docutils==0.20.1
requests==2.32.4
rstcheck==6.2.4
doc8==1.1.2
sphinx_sitemap==2.6.0
sphinx-last-updated-by-git==0.3.8
sphinx-autobuild==2024.10.3
# sphinx-external-toc
sphinx-design==0.6.0
pylint==3.3.6
myst-parser==4.0.1
pydata-sphinx-theme==0.16.1
# sphinxcontrib-bibtex
ablog==0.11.12
sphinx-reredirects==0.1.6
sphinx_design_elements==0.4.0
sphinx-tags
sphinxcontrib-mermaid
graphviz

# Typesense search integration (optional)
typesense==0.21.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1